# cmd/docs/swagger.yaml
basePath: /
definitions:
  AddPostRequest:
    properties:
      category:
        type: string
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  UpdatePostRequest:
    properties:
      category:
        type: string
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  handlers.AddPostResponse:
    properties:
      id:
        type: integer
    type: object
  models.Post:
    properties:
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple blog API
  title: Blog API
  version: "1.0"
paths:
  /posts:
    get:
      consumes:
      - application/json
      description: Get all posts
      operationId: get-all-posts
      parameters:
      - description: Search term
        in: query
        name: term
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: internal error
          schema:
            type: string
      summary: Get all posts
    post:
      consumes:
      - application/json
      description: Add post
      operationId: add-post
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/AddPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handlers.AddPostResponse'
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Add post
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete post by id
      operationId: delete-post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Delete post
    get:
      consumes:
      - application/json
      description: Get post by id
      operationId: get-post-by-id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Get post by id
    put:
      consumes:
      - application/json
      description: Update post by id
      operationId: update-post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/UpdatePostRequest'
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Update post
schemes:
- http
swagger: "2.0"